# src/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Dowin API
  description: Dowin(가중목) 서비스 API 명세서
  version: 0.1.0
servers:
  - url: /api
    description: Local development server
paths:
...
  /workspaces/{id}:
    put:
      summary: "[ADMIN] 워크스페이스 이름 변경"
      tags: [Workspace]
      security: [{ sessionCookie: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "새 워크스페이스 이름" }
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "403":
          description: 해당 워크스페이스 관리자 권한 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: 워크스페이스를 찾을 수 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: "[ADMIN] 워크스페이스 삭제"
      tags: [Workspace]
      security: [{ sessionCookie: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: 삭제 성공
        "403":
          description: 해당 워크스페이스 관리자 권한 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: 워크스페이스를 찾을 수 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  schemas:
...
    Workspace:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        planCode:
          type: string
          enum: [FREE, STANDARD]
        memberCount:
          type: integer
          minimum: 0
        freeMemberLimit:
          type: integer
          minimum: 1
        isOverFreeMemberLimit:
          type: boolean
        createdAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: 머신 리더블 에러 코드 }
            message:
              { type: string, description: 사람이 읽을 수 있는 한국어 메시지 }
            details:
              {
                type: object,
                description: 유효성 검사 실패 시 필드별 상세 오류 (선택),
              }
  responses:
    UnauthorizedError:
      description: 로그인이 필요합니다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dowin_sid
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: dowin_admin_sid